First_laba_try_33

+++ tasks/moskaev_v_max_value_elem_matrix/mpi/src/ops_mpi.cpp
@@ -46,7 +46,7 @@ bool MoskaevVMaxValueElemMatrixMPI::RunImpl() {
   // ИСПРАВЛЕНИЕ 1: Правильное распределение строк
   size_t start_row = rank * rows_per_process + std::min(static_cast<size_t>(rank), remainder);
   size_t end_row = start_row + rows_per_process;
-  if (rank < remainder) {
+  if (static_cast<size_t>(rank) < remainder) {
     end_row += 1;
   }
